factor out autograd-dependent tests
