+++ .github/workflows/build-ci.yml
@@ -55,6 +55,25 @@ jobs:
       - name: Test fmmax
         run: pytest tests/fmmax
 
+  test-grcwa:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+      - name: Setup python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.10"
+          cache: "pip"
+          cache-dependency-path: pyproject.toml
+      - name: Setup environment
+        run: |
+            python -m pip install --upgrade pip
+            pip install "numpy<2"
+            pip install ".[tests,dev]"
+      - name: Test fmmax
+        run: pytest tests/fmmax
+
   test-examples:
     runs-on: ubuntu-latest
     steps:
